Refactor SAT cipher model
The SAT cipher model has been refactored to be built as a 3-CNF. In
doing this, the variables' id has been refactored in order to have
labels consisting of 3 digits. This will allow consistent sorting.

The parsing of the solver output has been extended in cipher model
module. The same will be applied to other models.

Auxiliary funcions in models' utils have been refactored to be better
tailored for models.